pizza-flow-simulator-random/app/worker/requirements.txt
schedule
pydantic
requests
boto3
kubernetes==31.0.0
cryptography
python-socketio[asyncio_client]
